Limit access to the device's fixed 802.11 MAC address.
Changes getifaddrs() to return an anonymized MAC (02:00:00:00:00:00) for all interfaces if the process requesting it is a not system process (i.e. if it has a UID >= 10000).

Note that this change is not meant to provide any proper security
guarantees (indeed, it is part of the app's process, making it easy to
circumvent). Instead, we will use it to understand which apps are
actually calling the method (as opposed to just referencing it in their
code), and to see when apps are breaking across the tester population.

Bug: 141455849
Test: atest bionic-unit-tests-static
Test: Connect to Wi-Fi network
Test: Set up hotspot
Test: Call getifaddrs() directly or through NetworkInterface.
Change-Id: I4135b43989e135adc7272700d1dfdf0e87b75978