Update patch set 1
Patch Set 1:

This depends on not only my memory tagging patch but also the scudo switch back to scudo relanding, so mostly sending it out for an early look. The main thing that I'd like feedback on is whether this should be a mallopt or just a new function. Right now this just controls the heap but eventually it should also end up controlling the stack and globals so maybe it should be a separate function?

Patch-set: 1