Delete unnecessary (and sometimes problematic) attributes on libdl functions.
The weak attributes on declarations were causing ld-android.so to be dropped
from libdl.so's DT_NEEDED at link time when linked with --as-needed. Somehow
things kept working most of the time, but it caused some binaries to start
segfaulting at startup.

While here, drop the visibility attributes on the declarations (which are
redundant, declarations have default visibility by default) as well as the
weak attributes on most of the definitions (weak has no effect on defined
symbols at the dynamic loader level).

Bug: 146456667
Change-Id: I452af9a2b605bca3a8432fc40d112c83cbd858df