Update patch set 3
Patch Set 3:

> Patch Set 1:
> 
> This depends on not only my memory tagging patch but also the scudo switch back to scudo relanding, so mostly sending it out for an early look. The main thing that I'd like feedback on is whether this should be a mallopt or just a new function. Right now this just controls the heap but eventually it should also end up controlling the stack and globals so maybe it should be a separate function?

I think it should be a separate function, exactly because it does not just control malloc(). Stack tagging is not there yet, but I don't think we want to wait for it to land to have a new interface to control tag checking.

That said, as I commented on the Scudo change, I would rather have the tag checking mode be completely orthogonal to tagging being enabled or not in Scudo. Therefore it would make sense to keep a mallopt() to disable tagging in Scudo (presumably we only want to allow enabling when Scudo is initialised).

Patch-set: 3