RFC: platform profiler signal: add traced_perf codepath
This patch adds a case for the profiling signal handler (previously just
for native heapprofd profiling) when si_value == 1, corresponding to
traced_perf being the requesting party.

The handler opens /proc/self/{maps,mem}, connects to (init-created)
/dev/socket/traced_perf, and then sends the fds over the socket.
Everything happens synchronously within the signal handler. Socket is
made non-blocking, and we do not retry.

Please treat this as an RFC at this point. I'm looking for feedback on:
* approach/structure
* what is the most appropriate way to test this within bionic.

Otherwise, it's been tested manually.

DO NOT SUBMIT: needs traced_perf daemonization to be submitted, as the
socket doesn't exist at HEAD at the time of writing.

Bug: 144281346
Change-Id: Iea904694caeefe317ed8818e5b150e8819af91c2