Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 1:
> > 
> > This depends on not only my memory tagging patch but also the scudo switch back to scudo relanding, so mostly sending it out for an early look. The main thing that I'd like feedback on is whether this should be a mallopt or just a new function. Right now this just controls the heap but eventually it should also end up controlling the stack and globals so maybe it should be a separate function?
> 
> I think it should be a separate function, exactly because it does not just control malloc(). Stack tagging is not there yet, but I don't think we want to wait for it to land to have a new interface to control tag checking.
>
> That said, as I commented on the Scudo change, I would rather have the tag checking mode be completely orthogonal to tagging being enabled or not in Scudo. Therefore it would make sense to keep a mallopt() to disable tagging in Scudo (presumably we only want to allow enabling when Scudo is initialised).

In that case, the new function would just be a wrapper around prctl, right? Basically this, but with more error handling:

void set_memory_tag_level(int level) {
  int tagged_addr_ctrl = prctl(PR_GET_TAGGED_ADDR_CTRL, 0, 0, 0, 0);
  tagged_addr_ctrl = (tagged_addr_ctrl & ~PR_MTE_TCF_MASK) | level;
  prctl(PR_SET_TAGGED_ADDR_CTRL, tagged_addr_ctrl, 0, 0, 0);
}

If that's all that the function does, maybe the application should just do that itself? The mallopt to disable heap memory tagging sounds good otherwise.

I wouldn't add an interface for enabling memory tagging since we don't have a use case for it and it would introduce a certain amount of complexity. Instead, all processes would start up with memory tagging enabled and the process would need to disable it if necessary.

Patch-set: 3