Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I want to reduce DSO proliferation on the system. Why do we need a libunwind.so instead of just exporting the unwinder from libc?

I think we could export the unwinder from libc, although this would mean that the unwinder symbols would be exposed to the NDK (we'd be adding them to LIBC_PRIVATE so they wouldn't appear in the NDK stubs but could still be used at runtime for symbol resolution). I think this is unlikely to cause problems but I'd like to get opinions from Bionic folks as well.

Patch-set: 4