Update patch set 3
Patch Set 3:

I want to hold off on submitting this for a couple of weeks. We hope to re-enable scudo in the next few days, and it's a bad idea to make two changes to the memory allocator without some space to determine which might be causing a problem.

I'll try and do a review of the code by the end of the week (assuming the scudo re-enabling goes okay).

Patch-set: 3
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>