Extends ifaddrs for cases where the client cannot send RTM_GETLINK requests.
The previous implementation of getifaddrs() depended on RTM_GETLINK
requests being allowed, returning an error otherwise. This change makes getifaddrs() attempt to get all necessary
information from RTM_NEWADDR messages when RTM_NEWLINK messages are not available.

The code is functionally the same when RTM_GETLINK requests are allowed.
When RTM_GETLINK requests are denied, only interfaces that have a
network address are returned, and physical addresses for these
interfaces remain unset.

This code also updates getifaddrs() to behave as if RTM_GETLINK requests
are not allowed for non-system apps that have a target API level of R.
This code will be removed when kernel changes that enforce this are merged, and are purely meant to check for potential appcompat issues beforehand.

Bug: 141455849
Test: atest bionic-unit-tests-static
Test: Connect to Wi-Fi network
Test: Set up hotspot
Test: Cast from device
Test: Pair Bluetooth device
Test: Call getifaddrs() directly from within an app.
Test: Call NetworkInterface#getNetworkInterfaces() from within an app.
Test: Repeat above tests with an app that targets Android R.

Change-Id: Ia47e037d181ca5df6d9fdae19b405cabfafc6b0f