Update patch set 8
Patch Set 8: Code-Review+2

(1 comment)

We discussed the previous issues I had offline, and the only one remaining was my worry that setting the memory limit and getting the heapprofd signal at the time would cause issues.

The answer is no. The new code simply discards the heapprofd signal if the memory limit wins the race. I also wrote a simple test to verify this, and I was able to see some cases where heapprofd is enabled properly, and some where the signal is ignored.

Since the heapprofd folks are fine with this choice, I'm fine with it too.

Patch-set: 8
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2