Allow android_mallopt(M_SET_HEAP_TAGGING_LEVEL) to control scudo heap tagging.
The tag level may now be async, which is now the default on MTE enabled
hardware. When the tag level is set to none, memory tagging is disabled in
the allocator using the new API proposed in https://reviews.llvm.org/D70762 .

Rework pointer tagging to allow it to coexist with heap tagging.

Bug: 135772972
Change-Id: I847f5822a70913c446ed9ffa13792177bbfc96af