Update patch set 5
Patch Set 5:

(9 comments)

There are no tests for this, I know it's not enabled yet, but is there a way to test this when it's enabled?

I worry that this will fundamentally break a few features of the allocator and it's hard to see with this cl alone.

Patch-set: 5
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>