Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Presubmit-Ready+1
> > 
> > Should I add a link to the system core .clang-format-2 to replace this?
> 
> I think so.  Even if we can't run it in the presubmit hook because of the upstream code, it'll still be useful for people to run clang-format manually.

Adding in enh's statement (it didn't seem to get added by itself):
Should we just turn the standard one on, and say that anyone updating the upstream directories needs to use --no-verify? Which in practice should only ever affect me, because the whole point of those files is that there shouldn't be any *edits*, just wholesale replacements?



I would be in favor of this. In fact if you have a mix of files that should be formatted you could do the upload and revert the one you don't want to be formatted.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>