Update patch set 6
Patch Set 6:

(4 comments)

The Android.bp file is not correct, and doesn't build for me. That part is really complicated, so let me take a crack at fixing it up so it works with your stuff and works with the new scudo override stuff.

Patch-set: 6
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>