Update patch set 9
Patch Set 9:

(1 comment)

I finally got around to running some benchmarks to get a hang on how this affects performance, and it does affect it measurably.

On my walleye, the stdlib_malloc_free_decay1/64 benchmark increases:

On 32 bit: 205ns to 224ns
On 64 bit: 173ns to 185ns

Given this, I think you should use the hook mechanism the same way that perfetto does when gwp_asan is enabled.

This has the benefit that it guarantees that the limit functionality continues to work without any additional changes. It also incurs no performance penalty when it is disabled.

The downside is that perfetto and gwp_asan do not work together. I don't know if this is necessary.

Here's the command I used to get the data:

/data/benchmarktest64/bionic-benchmarks/bionic-benchmarks --benchmark_filter=BM_stdlib_malloc_free_decay1/64 --bionic_cpu=3 --benchmark_report_aggregates_only=true --benchmark_repetitions=1000

This locks to a cpu and runs it 1000 times to get as accurate timing as possible.

Patch-set: 9
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>