Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> (1 comment)
> 
> I finally got around to running some benchmarks to get a hang on how this affects performance, and it does affect it measurably.
> 
> On my walleye, the stdlib_malloc_free_decay1/64 benchmark increases:
> 
> On 32 bit: 205ns to 224ns
> On 64 bit: 173ns to 185ns
> 
> Given this, I think you should use the hook mechanism the same way that perfetto does when gwp_asan is enabled.
> 
> This has the benefit that it guarantees that the limit functionality continues to work without any additional changes. It also incurs no performance penalty when it is disabled.
> 
> The downside is that perfetto and gwp_asan do not work together. I don't know if this is necessary.

I think this is a great idea. Heapprofd code already handles replacement of malloc hooks in a zygote child. Since we will be profiling only a small subset of processes, this way we can have exactly zero overhead in ~99% of cases.

We will need to handle the case when perfetto attempts to take over from gwp-asan, because that just won't work - the underlying allocator won't be able to deallocate a gwp-asan pointer. It makes sense to chain the allocator the same way malloc_limit does, but we also (ideally) should be compatible with malloc_limit.

We could add a third level of malloc dispatch in addition to current_dispatch_table and default_dispatch_table. Gwp-asan would go below both malloc_limit and perfetto, so we'd need to teach perfetto to forward to gwp-asan the same way that malloc_limit forwards to perfetto. How does that sound?

Patch-set: 9
Reviewer: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>