Update patch set 9
Patch Set 9:

> Given this, I think you should use the hook mechanism the same way that perfetto does when gwp_asan is enabled.

Turns out responding to the email doesn't add anything to the comments, my thoughts below

----

Given that we intend on enabling GWP-ASan for a runtime-sampled set of processes, this sounds like a good idea. It would be beneficial to introduce GWP-ASan directly into the dispatch table to avoid an extra layer of indirection overhead for a sampled process. Caveat: We still need GWP-ASan statically linked into libc so it's part of the initial set of modules, as we have IE TLS variables. 

I see no problems with incompatibilities with perfetto (assuming it uses the MallocHooks). Similar to heapprofd or malloc_debug, we'd just never install GWP-ASan when these libraries are used.

Patch-set: 9