Ignore symbols of imported libs' dependencies
When a library is present in a namespace via the secondary_namespaces list
(i.e. the executable, LD_PRELOAD, DF_1_GLOBAL, or android_create_namespace
inheritance), then we want to search that library's symbols, but not the
symbols of its dependencies. Otherwise, we want to search the dependencies
to handle cross-NS dependency.

Bug: http://b/148569846
Test: bionic unit tests
Change-Id: If798d69de28ed5c0f1a155e4ff85c7e08934e531