Update patch set 6
Patch Set 6:

That should just about do the trick - we specifically don't enable GWP-ASan if heapprofd/etc are enabled iff we aren't the first one installed. If GWP-ASan chooses to sample a process in libc, we are guaranteed to be first, but if we are turning on in a zygote child, GWP-ASan can lose the race. If this is the case, we simply don't install GWP-ASan, as this is expected to be rare and heapprofd is debug-only.

Let me cherrypick onto an internal branch and run tests.

Patch-set: 6