Update patch set 7
Patch Set 7:

> Patch Set 7:
> You need to follow the perfetto model of setting default and then setting current if it's not already nullptr, and everything should just work.

Yep, this is what we do right now. The only addition is that for malloc_hooks/malloc_debug/heapprofd - if they initialise after GWP-ASan, they get provided GWP-ASan as their dispatch pointer to _initialize() instead of je_malloc (__libc_malloc_default_dispatch).

>I think only enabling one of gwp_asan and perfetto at the same time seems reasonable to me, so no need for yet another dispatch table.

The current sol'n allows for GWP-ASan and perfetto if GWP-ASan is turned on first, with a graceful fallback to perfetto-only if it's first. Is this OK?

Patch-set: 7