Parse ELF notes in linker to determine if zygote.
We want to be able to determine if a process is a zygote. The zygote
includes an ELF note, and we parse the ELF notes in the linker to find
if it's present. If it is -> we set a global that's shared between libc
and the linker.

This also includes the requisite plumbing for the .shstrtab section.
This previously was ignored by the linker, and this patch adds support
for parsing more ELF notes.

Bug: 135634846
Test: check app_process[32/64] has zygote true, else false.
Change-Id: I94d195a268627f165eb9797ae970998541118788