Update patch set 3
Patch Set 3:

(5 comments)

> Patch Set 1:
> 
> (5 comments)
> 
> eugenis/rprichard: how certain are we about this API? obviously we can't retcon this into API 29, but i had assumed we were actually thinking about 3*1* so we have a year to realize we've screwed up and fix it, rather than jamming it in to R (30) at the last minute.
> 
> (for 31 i have no reservations at all --- but 30 is very close to the last minute at this point.)
> 

Happy to change it to 31. When is the absolute cut-off for 30?

> ((i'll skip my usual "where's the test?" since there's obviously a chicken-and-egg situation going on here!))

What's the best way to add a test for this?
What I've done to test this was manual. (ie., build bionic + llvm, then push to the device and run a binary to confirm it 'works')

Patch-set: 3