Update patch set 9
Patch Set 9:

(2 comments)

I'm fine with this in general, given that we have at least a year to change our minds, and  the potential API for dynamic TLS seems likely to be in addition to this, and not instead of it.

Should the function name start with "android_" instead of "__libc_"?

Patch-set: 9
Reviewer: Gerrit User 1005967 <1005967@85c56323-6fa9-3386-8a01-6480fb634889>