Update patch set 4
Patch Set 4:

> Patch Set 4: Autosubmit+1 Presubmit-Ready+1
> 
> (2 comments)
> 
> > Patch Set 3: Code-Review+2
> > 
> > (1 comment)
> > 
> > Thanks a lot. LGTM with a minor comment (on top of Elliott's).
> > I don't see what can possibly go wrong with a signal handler*. Ship it ;)
> > 
> > * From my personal collection: https://crbug.com/448968, https://crbug.com/483399, https://crbug.com/481420, https://crbug.com/477444, https://crbug.com/473973, https://crbug.com/249880
> 
> I'm morbidly curious about what will cause us to end up on a postmortem with this.

See the good side. Between Elliott and myself, we probably saw the worst horror stories of signal handlers. If something else that missed the review happens, it's guaranteed to be a tantalising and spectacularly unique gem for the collection.
Something like: if you are in two nested signal handlers and within a seecomp-ed process, in a night of full moon of a leap year the kernel corrupts a register while restoring the context.

Patch-set: 4
Reviewer: Gerrit User 1040443 <1040443@85c56323-6fa9-3386-8a01-6480fb634889>