Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Do the perfetto tests cover all three cases?
> 
> It doesn't. We weren't even aware that these three cases exist until we found the crashes.
> 
> mitchp@: Is there some way we could control where GWP-ASAN is on in the target process? Then we could exercise these three cases.

Explicitly testing - no. In practice, with enough runs, you get GWP-ASan everywhere.

We do expose an android_mallopt() for turning GWP-ASan on, but that comes with a whole bunch of restrictions. It has to be done in a single-threaded context, and you can't turn GWP-ASan *off* after it's been turned on. Furthermore, the only reliable way to test whether GWP-ASan is on is to scan /proc/maps.

For now, are we okay with fixing b/150085813 with this patch, and I'll have a think about the story for testing this more reliably?

Patch-set: 3
Reviewer: Gerrit User 1453063 <1453063@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0