wrap execve, to ignore out-of-process platform signals before syscall
There are normally custom handler functions attached to these
platform-internal signals. If a process calls execve, the kernel has
to reset the signals' dispositions back to the default (SIG_DFL), which
is to terminate the process. This is problematic for signals that are
delivered from a separate process, as the latter cannot guarantee that
the (innately asynchronous) signal won't be delivered right after an
execve in the target process.

By setting the signals' dispositions to SIG_IGN right before the execve
syscall, we safeguard the newly installed process image from being
inadvertently terminated. This is due to the kernel retaining the
ignored status across the execve. For a typical process, the debugger
signal handling will be reinstated in the dynamic linker, and the
profiling signal will be handled by libc's initialization.

Separately, this effectively changes the stable state of:
* BIONIC_SIGNAL_PROFILER to SIG_IGN in static libc builds (which don't
  define a handler).
* BIONIC_SIGNAL_ART_PROFILER to SIG_IGN in most processes,
  as there is no builtin handler for it in libc itself (it's
  conditionally overriden by ART in managed app processes).

Both of these properties are desireable from the profilers' side (and
can be seen as an independent bugfix). I wanted to highlight that it is
an intentional change in behaviour that is not directly related to the
post-execve vulnerable window.

Parts I wasn't sure of:
* which of the reserved signals to ignore. I can speak for the profiler
  signals, and assumed that debuggerd should also be technically
  included due to the remote backtrace option. Please advise if you
  think more/fewer signals should be in this list.
* how to run bionic unittests with glibc.

I have not done any specific performance tests, I'm assuming that
several more syscalls prior to execve aren't going to be significant,
given that the syscall itself is quite heavy.

Tested: bionic 64 bit unittests (static and dynamic)
Tested: booted a crosshatch-userdebug (though it wasn't pure AOSP, I
        patched this onto an internal branch that is reasonably new).

Bug: 151328035
Bug: 151835887
Change-Id: If4d9eac86eaa9ebee69abd8c15336bf5174aefa1