Ignore ART profiling signal.
These solves the issue that targetting non-profilable apps crashed them.
There is still a race condition between starting the app and the SIG_IGN
being installed, but that will be fixed in follow-ups.

This also does not cover statically linked programs, which should not be
a concern for Zygote children / apps.

Bug: 151328035
Test: java profile sysui, doesn't crash

Change-Id: I6b99352ed50afe15a609f7ddb85312c2676ddf11