Create linker_basename / linker_dirname
There should be no functional change. The linker requires the GNU
(i.e. non-POSIX) basename behavior, but there's no good way to select
the GNU function, so just implement it in linker_utils.cpp.

The linker already had a dirname function that is similar to dirname
in libc, but:
 - it returns an std::string (not a TLS buffer or the input string)
 - it matches GNU basename's semantics for "/foo/" rather than POSIX

Matching the libc names is nice because it makes it harder to
unintentionally use the libc functions, but it's confusing and could
break if a file included libgen.h.

Bug: none
Test: bionic tests
Change-Id: I6b53b0c93557869da079d4a121af8cd246c37217