Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review+2
> 
> hmm... maybe the real bug is that libasync_safe shouldn't be a separate library at all, and should just be part of libc?
> 
> but that's not your fault, and this looks like an improvement on what's currently checked in, so +2...

Thanks for approving. I don't know the origin of libasync_safe, but given that it has some knowledge about the socket protocol to logd (which isn't guaranteed to be stable I guess), we don't want the lib be included in unbundled modules like APEX and want the lib be served from the runtime APEX only.

Fortunately, the runtime APEX is not updatable in R. So we have time to fix this. Filed b/152466437

Patch-set: 6