wrap execve, block reserved platform signals across the syscall
There are normally custom handler functions attached to these
platform-internal signals. If a process calls execve, the kernel has
to reset the signals' dispositions back to the default (SIG_DFL), which
is to terminate the process. This is problematic for signals that are
delivered from a separate process, as the latter cannot guarantee that
the (innately asynchronous) signal won't be delivered right after an
execve in the target process.

Therefore, we wrap the execve syscall in libc, blocking the reserved set
of signals (which is preserved across the syscall). We also modify the
system linker to unblock these signals early in its setup. Prior to
unblocking, we ignore (SIG_IGN) the subset of signals that we anticipate
to be left pending across the exec (or which could arrive early in the
new image's lifetime).

It is worth highlighting that pending/"early" signals in this subset
will therefore be dropped. This is done intentionally, as to not
complicate the design further. None of the signals from this patchset
strictly care about not being lost in corner cases (but we definitely
don't want to terminate processes).

The debugger signal handling will be reinstalled immediately afterwards,
in the linker. Profiling handler will be installed by libc's preinit
(unless the process is linking in the static version of bionic). The ART
profiling handler will be optionally installed by ART in managed
processes (and left ignored elsewhere).

Limitations:
* breaks POSIX by making execve not async-signal-safe.
* presumably, PT_INTERP in the ELF could make an exec'd process avoid
  the system linker. Even if this is possible in practice on Android,
  the outcome is that the signals stay blocked, and might be therefore
  queued by the kernel. The signalling rates should be low enough to not
  make this too bad in practice, even if it does happen.
* I'm pretty sure there are still plenty of other edge-cases around
  reserved signal handlers (thinking vforks and posix_spawn). It's
  unlikely that we can make them fully transparent.

Parts I wasn't sure of:
* which of the reserved signals to ignore. I can speak for the profiler
  signals, and assumed that debuggerd should also be technically
  included due to the remote backtrace option. Please advise if you
  think more/fewer signals should be in this list.

Tested: bionic 64 bit unittests (dynamic, I'm not sure what the right
        setup for static testing is)
Tested: booted a crosshatch-userdebug (though it wasn't pure AOSP, I
        patched this onto an internal branch that is reasonably new).

Bug: 151835887
Change-Id: If4d9eac86eaa9ebee69abd8c15336bf5174aefa1