Update patch set 1
Patch Set 1: Code-Review+2

> Patch Set 1:
> 
> This is a bit of a weakness, but hopefully not too severe, given that each process already has multiple unchecked shared libraries with writable segments that can be used to inject fake vtables.

this is only true for svelte devices anyway, right? (judging by the fact that the bug is the jemalloc VMA bug.)

worth clarifying the checkin comment slightly to make it explicit that scudo isn't affected by this, but that the svelte configuration of jemalloc is?

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2