Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> ifunc resolvers are very limited in what they can do, because they run in an unrelocated (under-relocated?) binary. I don't think we can reliably initialize hwasan in that context.

Mm, you may be right. I think we can't even rely on the GOT entry for __hwasan_init being initialized at this point. If we want to do something automatic here maybe all that we can do is disable hwasan instrumentation on functions that are marked as ifunc resolvers. The functions wouldn't be able to call other functions safely, but to some extent that's a pre-existing restriction of ifunc resolvers (since e.g. the function call could go via the PLT in some cases).

Patch-set: 3