Update patch set 11
Patch Set 11:

(5 comments)

does this mprotect() implementation mean that ART and mono (and v8?) "just work"? it seems like the ART changes now are just an optimization rather than actually necessary for correctness?

we'd want to get the app compat folks to sign off, but if it does "just work" for the three obvious likely-to-be-affected cases (and the cost to mprotect() users is small enough --- i don't want you back in a couple of months wanting to remove all our mprotect() calls :-) ) then i think we could give this a go.

Patch-set: 11
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>