Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Unfortunately old kernels don't correctly check for unknown open flags, so we don't have a reliable way to test for this. In terms of upgrading devices, that's no longer an issue, but we still have libc.a in the NDK. O_TMPFILE is on my list of reasons I'd like to drop support for libc.a...
> 
> 	/*
> 	 * In order to ensure programs get explicit errors when trying to use
> 	 * O_TMPFILE on old kernels, O_TMPFILE is implemented such that it
> 	 * looks like (O_DIRECTORY|O_RDWR & ~O_CREAT) to old kernels. But we
> 	 * have to require userspace to explicitly set it.
> 	 */
> 
> Also, in the future, we can try openat2(2), which always checks all its flags.

oh, good, that's handy. but since we still ship libc.a to run on JellyBean, the fallback code isn't going anywhere for several more years even if we change tmpfile() to use O_TMPFILE today.

(plus that's only 1 out of the 8 uses of ScopedSignalBlocker.)

so i'm assuming rprichard's "this seems like the right thing to do" applies to the "let's fix ScopedSignalBlocker" rather than any of the other things that have come up on this bug. (though we should do the other things too!)

Patch-set: 2