Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> can we build the tests as part of the libc tests instead?

We could (and I got most of the way towards implementing it), but one issue is that the tests use symbol interposition to change the behaviour of the allocator:

https://cs.android.com/android/platform/superproject/+/master:external/scudo/standalone/tests/scudo_unit_test_main.cpp;l=24

Which would mean that this behaviour change would apply to the other bionic tests as well, so we wouldn't be testing what we ship exactly. It doesn't appear to be a problem right now, but it could affect us in the future.

Patch-set: 1