Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> we could still have it as part of the bionic/tests/ build, even if it's not specifically in the same binary...

The only difference then would be that we'd move the location of the build rule? I wouldn't have thought that would be significant.

IIRC the reason why the bionic tests can link against the private symbols is the "bootstrap: true". And indeed, if I add "bootstrap: true" to scudo_unit_tests it builds without this change. So maybe we should just do that?

Patch-set: 1