Fix types of PAGE_SIZE, PAGE_MASK.
Previously, PAGE_SIZE was an unsuffixed integer literal, which resulted
in int being selected as its type. This results in code using PAGE_MASK
(or reimplementing PAGE_MASK themselves) clearing the top 32 bits of
pointers when running on 64-bit.

Switch the definition to a long, and while we're at it, make it unsigned
to match glibc.

Bug: http://b/155136053
Test: treehugger
Change-Id: I2e293d81c4aa44be172fc6c1a095098122b14011